Resolve deprecation and other warnings (#3444)
* Replace deprecated JavaConverters

* Remove unused/unneeded imports

* Supply empty argument list

* Convert to double explicitly

* Use methods directly on Either

* Replace .mapValues with .view.mapValues

* Call toIndexedSeq explicitly

* Avoid Stream

* Replace Array with Seq

* Replace replaceAllLiterally with replace

* Replace toIterable with toSeq

* Write .apply explicitly

* Replace fields with properties

+++ elastic4s-client-akka/src/main/scala/com/sksamuel/elastic4s/akka/DefaultBlacklist.scala
@@ -2,7 +2,7 @@ package com.sksamuel.elastic4s.akka
 
 import java.util.concurrent.ConcurrentHashMap
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 import scala.concurrent.duration.FiniteDuration
 
 /** Thread-safe host blacklist. Blacklist duration starts with `min` and exponentially increased up to `max` on
@@ -34,7 +34,7 @@ private[akka] class DefaultBlacklist(min: FiniteDuration, max: FiniteDuration, n
       val retries = record.retries + 1
 
       val untilTime = now + Math
-        .min(min.toNanos * Math.pow(2, retries * 0.5), max.toNanos)
+        .min(min.toNanos * Math.pow(2, retries * 0.5), max.toNanos.toDouble)
         .toLong
 
       val updatedRecord =
